test: fix a signed shift in test/string, and let ASan reach malloc's OOM path
Two failures in the native sanitizer build (tools/asan-build.sh):

test/string.c:119 built the top-bit-only pattern as 1LL << 63.  The left
operand is signed, so shifting into the sign bit is undefined (C99
6.5.7p4) and UBSan says so.  Build the value as 1ULL << 63 and convert to
long long; ffsll() still sees the same bit pattern, so the assertion is
unchanged.  It is the only shift in the file -- ffs(0x80000000) is fine,
a hex constant that large is already unsigned int.

test/malloc.c asserts that malloc() returns NULL with ENOMEM for a
request that cannot be satisfied, which is what C99 7.20.3.3p3 requires.
ntlibc gets that right: it passes the size through untouched (ASan
reports 0xffffffffffffff9b for malloc(SIZE_MAX - 100) -- no overflow on
the way in), so this is not a library bug.  It is ASan's default
allocator_may_return_null=0, which aborts inside ASan's own allocator
before the library can return anything.  Setting the option makes ASan
behave like a conforming allocator; no assertion is relaxed.

The option is set for that one test only, so every other sanitized test
keeps the strict default and a genuinely bogus huge allocation elsewhere
still aborts.  test/malloc.c defines __asan_default_options() so the
setting travels with the binary, and asan-build.sh also passes it in the
environment because the dynamic runtime the script needs
(-shared-libasan) does not let a program's definition preempt the one
inside libclang_rt.asan.so.

asan: 4/7 -> 6/7 (test/stdio still fails on an unrelated LLONG_MIN
negation in src/stdio/scanf.c).  make check stays 15/15 on x86_64 and
i386; make lint unchanged.

+++ test/string.c
@@ -116,7 +116,7 @@ int main(void)
 	strcpy(buf, "hello"); CHECK(index(buf, 'l') == buf + 2 && rindex(buf, 'l') == buf + 3);
 	CHECK(ffs(0) == 0 && ffs(1) == 1 && ffs(0x80) == 8 && ffs(-1) == 1 && ffs(0x80000000) == 32);
 	CHECK(ffsl(0) == 0 && ffsl(0x100) == 9 && ffsl(-1L) == 1);
-	CHECK(ffsll(0) == 0 && ffsll(1LL << 40) == 41 && ffsll(-1LL) == 1 && ffsll(1LL << 63) == 64);
+	CHECK(ffsll(0) == 0 && ffsll(1LL << 40) == 41 && ffsll(-1LL) == 1 && ffsll((long long)(1ULL << 63)) == 64);
 
 	/* errors and signals */
 	CHECK(streq(strerror(0), "No error information"));
